requirements.txt
django-ckeditor
django-select2
django-debug-toolbar